changes after meerge

+++ web Drinky-Winky/src/main/java/com/infoshare/service/DrinkService.java
@@ -8,8 +8,10 @@ import org.slf4j.LoggerFactory;
 
 import javax.ejb.EJB;
 import javax.enterprise.context.RequestScoped;
+import javax.inject.Inject;
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Transactional
@@ -19,17 +21,24 @@ public class DrinkService {
 
     @EJB
     DrinkDao drinkDao;
+    @Inject
+    MessageService messageService;
 
-    public void addDrink(DrinkDTO drinkDTO) {
-        Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
-        drinkDao.addDrink(drink);
+    public DrinkDTO addDrink(DrinkDTO drinkDTO) {
+        if (drinkDao.getDrinkList().stream().noneMatch(drink -> drink.getName().equals(drinkDTO.getName()))) {
+            Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
+            messageService.leaveMessage(1L, "Drink was added!");
+            return DrinkDTO.drinkToDTO(drinkDao.addDrink(drink));
+        } else {
+            messageService.leaveMessage(1L, "Nothing has happened, please change name.");
+            return null;
+        }
     }
 
     public void editDrink(DrinkDTO drinkDTO) {
         Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
         drinkDao.editDrink(drink);
     }
-
 
     public DrinkDTO getDrinkByName(String name) {
         Drink drinkByName = drinkDao.getDrinkByName(name);
@@ -39,7 +48,6 @@ public class DrinkService {
         return null;
     }
 
-
     public DrinkDTO getDrinkById(Long id) {
         Drink drinkById = drinkDao.getDrinkById(id);
         if (drinkById != null) {
@@ -48,12 +56,12 @@ public class DrinkService {
         return null;
     }
 
-
     public void deleteDrinkById(Long id) {
-        drinkDao.deleteDrinkById(id);
-        logger.info("Drink has been deleted");
+        if (drinkDao.getDrinkList().stream().anyMatch(drink -> drink.getId().equals(id))) {
+            drinkDao.deleteDrinkById(id);
+            logger.info("Drink has been deleted");
+        }
     }
-
 
     public void deleteDrinkByName(String name) {
         if (drinkDao.getDrinkList().stream().anyMatch(drink -> drink.getName().equals(name))) {
@@ -91,7 +99,6 @@ public class DrinkService {
         return getDrinkList().subList(fromIndex, toIndex);
     }
 
-
     public Set<String> getUniqueGlassesNameList() {
         return drinkDao.getDrinkList()
                 .stream()
@@ -99,6 +106,75 @@ public class DrinkService {
                 .collect(Collectors.toSet());
     }
 
+    public List<DrinkDTO> findDrinkListByIngredient(String ingredient) {
+        if (ingredient != null || !ingredient.isBlank()) {
+            List<DrinkDTO> found = new ArrayList<>();
+            List<DrinkDTO> drinkList = getDrinkList();
+            for (DrinkDTO foundDrink : drinkList) {
+                if (foundDrink.getIngredientList().stream().anyMatch(d -> d.getName().toLowerCase().contains(ingredient.toLowerCase()))) {
+                    found.add(foundDrink);
+                }
+            }
+            return found;
+        }
+        return null;
+    }
+
+    public List<DrinkDTO> findDrinkListByType(String type) {
+        if (type != null || !type.isBlank()) {
+            List<DrinkDTO> found = new ArrayList<>();
+            List<DrinkDTO> drinkList = getDrinkList();
+            for (DrinkDTO foundDrink : drinkList) {
+                if (foundDrink.getDrinkType().toLowerCase().contains(type.toLowerCase())) {
+                    found.add(foundDrink);
+                }
+            }
+            return found;
+        }
+        return null;
+    }
+
+    public List<DrinkDTO> findDrinkListBasicOnAllValues(String input) {
+
+        if (input != null || !input.isBlank()) {
+            List<DrinkDTO> found = new ArrayList<>();
+            List<DrinkDTO> drinkList = getDrinkList();
+            for (DrinkDTO foundDrink : drinkList) {
+
+                //Name check
+                if (foundDrink.getName().toLowerCase().contains(input.toLowerCase())) {
+                    found.add(foundDrink);
+                }
+                //glass type check
+                if (foundDrink.getGlassType().toLowerCase().contains(input.toLowerCase())) {
+                    if (!found.contains(foundDrink)) {
+                        found.add(foundDrink);
+                    }
+                }
+                //drink type check, if user want to find alcoholic or non alcoholic
+                if (foundDrink.getDrinkType().toLowerCase().contains(input.toLowerCase())) {
+                    if (!found.contains(foundDrink)) {
+                        found.add(foundDrink);
+                    }
+                }
+                // category check
+                if (foundDrink.getCategory().toString().toLowerCase().contains(input.toLowerCase())) {
+                    if (!found.contains(foundDrink)) {
+                        found.add(foundDrink);
+                    }
+                }
+                // trying to find an ingredient
+                if (foundDrink.getIngredientList().stream().anyMatch(d -> d.getName().toLowerCase().contains(input.toLowerCase()))) {
+                    if (!found.contains(foundDrink)) {
+                        found.add(foundDrink);
+                    }
+                }
+            }
+            return found;
+        }
+        return null;
+    }
+}
     public List<Drink> findRecipeByCategoryId(List<Long> ids) {
         return drinkDao.findDrinkByCategoryId(ids);
     }
